+++ packages/user/src/components/layout/index.tsx
@@ -1,9 +1,9 @@
 import { Outlet } from 'react-router-dom';
-import BodyContainer from 'src/components/layout/BodyContainer';
-import Banner from 'src/components/layout/top/Banner';
-import Footer from 'src/components/layout/top/footer';
-import Header from 'src/components/layout/top/header';
-import TopSectionContainer from 'src/components/layout/top/TopSectionContainer';
+import Banner from './banner';
+import BodyContainer from './BodyContainer';
+import Footer from './footer';
+import Header from './header';
+import TopSectionContainer from './TopSectionContainer';
 
 export default function Layout() {
 	return (
